Refactor: reorganize directories for client and server
