+++ packages/experiments/demo/client/demo.ts
@@ -1,4 +1,4 @@
-import { createView } from './demoView';
+import { createView } from '../demoView';
 import 'codemirror-6/codemirror.next/view/style/editorview.css';
 import 'codemirror-theme-ubuntu/codemirror-ubuntu-theme.css';
 import './styles.css';
