+++ packages/experiments/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "watch-demo": "rollup -w -c rollup.config.js",
-    "start": "node -r ts-node/register demo/server.ts",
+    "start": "node -r ts-node/register demo/server/server.ts",
     "test": "mocha"
   },
   "repository": {
@@ -31,12 +31,15 @@
     "typescript": "^3.1.3"
   },
   "dependencies": {
+    "@teamwork/websocket-json-stream": "^1.1.0",
     "@types/express": "^4.16.0",
     "@types/jsdom": "^12.2.0",
     "codemirror-6": "1",
     "codemirror-ot": "1",
     "codemirror-theme-ubuntu": "1",
     "express": "^4.16.4",
-    "jsdom": "^13.0.0"
+    "jsdom": "^13.0.0",
+    "sharedb": "^0.11.37",
+    "ws": "^6.1.0"
   }
 }
