+++ packages/experiments/rollup.config.js
@@ -4,7 +4,7 @@ import nodeResolve from 'rollup-plugin-node-resolve'
 import postcss from 'rollup-plugin-postcss'
 
 export default {
-  input: './demo/demo.ts',
+  input: './demo/client/demo.ts',
   output: {
     format: 'umd',
     file: './demo/build/demo_built.js',
